requirements.txt
llama-index==0.12.7
llama-index-embeddings-huggingface==0.4.0
llama-index-llms-openai==0.3.12
pandas==2.2.3
ipywidgets==8.1.5
llama-index-readers-file==0.4.1
pymupdf
matplotlib==3.10.0
gradio==5.9.1
llama-index-vector-stores-faiss==0.3.0
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1